requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
icalendar==5.0.10
idna==3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.0.6
